🧑‍💻 chore: add supporting features and enhancements
Backend enhancements:
- Add PATCH /transactions/{id} endpoint for updating transaction category
- Add category suggestion endpoints with AI-powered inference
- Enhance transaction API with better response models and data loading
- Improve parser detector and service layer logic
- Update schemas with new transaction and categorization fields

Frontend enhancements:
- Enhance FileUpload component with better error handling and feedback
- Improve StatCard styling and layout
- Add settings and configuration improvements

Misc:
- Update launch.json configuration
- Add WebullPDF parser for PDF statement parsing
- Add test files for comprehensive testing

+++ backend/app/services/market_data.py
@@ -93,10 +93,10 @@ def fetch_historical_prices_sync(symbol: str, start_date: date, end_date: date =
         prices.append(
             {
                 "price_date": idx.date(),
-                "open_price": round(row["Open"], 4),
-                "high_price": round(row["High"], 4),
-                "low_price": round(row["Low"], 4),
-                "close_price": round(row["Close"], 4),
+                "open_price": float(round(row["Open"], 4)),
+                "high_price": float(round(row["High"], 4)),
+                "low_price": float(round(row["Low"], 4)),
+                "close_price": float(round(row["Close"], 4)),
                 "volume": int(row["Volume"]) if row["Volume"] else None,
             }
         )
